refactor(xtream): extract request helpers into xtream-request.util
Move formatXtreamError, createXtreamErrorResponse and buildXtreamApiUrl
out of xtream.events.ts into a focused xtream-request.util.ts. Behavior
is unchanged; xtream.events.ts now holds only IPC registration and drops
from 380 to 273 lines, back under the repo's 300-line guideline
(CLAUDE.md TypeScript File Size Rule).

Verified: electron-backend build + lint clean, xtream.events.spec green
(372 passed; the 13 failures are the pre-existing Windows-only platform
suites unrelated to this change).
